+++ ecs-20140526/src/Models/CreateLaunchTemplateVersionRequest.php
@@ -8,6 +8,7 @@ use AlibabaCloud\Dara\Model;
 use AlibabaCloud\SDK\Ecs\V20140526\Models\CreateLaunchTemplateVersionRequest\dataDisk;
 use AlibabaCloud\SDK\Ecs\V20140526\Models\CreateLaunchTemplateVersionRequest\imageOptions;
 use AlibabaCloud\SDK\Ecs\V20140526\Models\CreateLaunchTemplateVersionRequest\networkInterface;
+use AlibabaCloud\SDK\Ecs\V20140526\Models\CreateLaunchTemplateVersionRequest\securityOptions;
 use AlibabaCloud\SDK\Ecs\V20140526\Models\CreateLaunchTemplateVersionRequest\systemDisk;
 use AlibabaCloud\SDK\Ecs\V20140526\Models\CreateLaunchTemplateVersionRequest\tag;
 
@@ -234,6 +235,11 @@ class CreateLaunchTemplateVersionRequest extends Model
     public $securityGroupIds;
 
     /**
+     * @var securityOptions
+     */
+    public $securityOptions;
+
+    /**
      * @var int
      */
     public $spotDuration;
@@ -322,6 +328,7 @@ class CreateLaunchTemplateVersionRequest extends Model
         'securityEnhancementStrategy' => 'SecurityEnhancementStrategy',
         'securityGroupId' => 'SecurityGroupId',
         'securityGroupIds' => 'SecurityGroupIds',
+        'securityOptions' => 'SecurityOptions',
         'spotDuration' => 'SpotDuration',
         'spotPriceLimit' => 'SpotPriceLimit',
         'spotStrategy' => 'SpotStrategy',
@@ -349,6 +356,9 @@ class CreateLaunchTemplateVersionRequest extends Model
         }
         if (\is_array($this->securityGroupIds)) {
             Model::validateArray($this->securityGroupIds);
+        }
+        if (null !== $this->securityOptions) {
+            $this->securityOptions->validate();
         }
         if (\is_array($this->tag)) {
             Model::validateArray($this->tag);
@@ -554,6 +564,10 @@ class CreateLaunchTemplateVersionRequest extends Model
                     ++$n1;
                 }
             }
+        }
+
+        if (null !== $this->securityOptions) {
+            $res['SecurityOptions'] = null !== $this->securityOptions ? $this->securityOptions->toArray($noStream) : $this->securityOptions;
         }
 
         if (null !== $this->spotDuration) {
@@ -805,6 +819,10 @@ class CreateLaunchTemplateVersionRequest extends Model
                     ++$n1;
                 }
             }
+        }
+
+        if (isset($map['SecurityOptions'])) {
+            $model->securityOptions = securityOptions::fromMap($map['SecurityOptions']);
         }
 
         if (isset($map['SpotDuration'])) {