+++ ecs-20140526/src/Models/StartTerminalSessionRequest/encryptionOptions.php
@@ -1,0 +1,76 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Ecs\V20140526\Models\StartTerminalSessionRequest;
+
+use AlibabaCloud\Dara\Model;
+
+class encryptionOptions extends Model
+{
+    /**
+     * @var bool
+     */
+    public $enabled;
+
+    /**
+     * @var string
+     */
+    public $KMSKeyId;
+
+    /**
+     * @var string
+     */
+    public $mode;
+    protected $_name = [
+        'enabled' => 'Enabled',
+        'KMSKeyId' => 'KMSKeyId',
+        'mode' => 'Mode',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->enabled) {
+            $res['Enabled'] = $this->enabled;
+        }
+
+        if (null !== $this->KMSKeyId) {
+            $res['KMSKeyId'] = $this->KMSKeyId;
+        }
+
+        if (null !== $this->mode) {
+            $res['Mode'] = $this->mode;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['Enabled'])) {
+            $model->enabled = $map['Enabled'];
+        }
+
+        if (isset($map['KMSKeyId'])) {
+            $model->KMSKeyId = $map['KMSKeyId'];
+        }
+
+        if (isset($map['Mode'])) {
+            $model->mode = $map['Mode'];
+        }
+
+        return $model;
+    }
+}